Feature 3 complete: Global.asax.cs application initialization conversion to .NET Core Program.cs. Starting Feature 4: Configuration management conversion
